+++ 1.20/Common/src/main/java/fuzs/universalbonemeal/UniversalBoneMeal.java
@@ -1,6 +1,5 @@
 package fuzs.universalbonemeal;
 
-import com.google.common.collect.Sets;
 import fuzs.puzzleslib.api.config.v3.ConfigHolder;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
 import fuzs.puzzleslib.api.core.v1.context.ModLifecycleContext;
@@ -8,10 +7,13 @@ import fuzs.puzzleslib.api.event.v1.entity.player.BonemealCallback;
 import fuzs.universalbonemeal.config.ServerConfig;
 import fuzs.universalbonemeal.handler.BonemealHandler;
 import fuzs.universalbonemeal.world.level.block.behavior.*;
+import net.minecraft.core.Direction;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.world.level.block.Blocks;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.Set;
 
 public class UniversalBoneMeal implements ModConstructor {
     public static final String MOD_ID = "universalbonemeal";
@@ -40,15 +42,16 @@ public class UniversalBoneMeal implements ModConstructor {
         BonemealHandler.registerBehavior(Blocks.SUGAR_CANE, SimpleGrowingPlantBehavior::new, () -> CONFIG.get(ServerConfig.class).allowSugarCane);
         BonemealHandler.registerBehavior(Blocks.VINE, VineBehavior::new, () -> CONFIG.get(ServerConfig.class).allowVines);
         BonemealHandler.registerBehavior(Blocks.NETHER_WART, NetherWartBehavior::new, () -> CONFIG.get(ServerConfig.class).allowNetherWart);
-        BonemealHandler.registerBehavior(Sets.newHashSet(Blocks.MELON_STEM, Blocks.PUMPKIN_STEM), FruitStemBehavior::new, () -> CONFIG.get(ServerConfig.class).allowFruitStems);
+        BonemealHandler.registerBehavior(Set.of(Blocks.MELON_STEM, Blocks.PUMPKIN_STEM), FruitStemBehavior::new, () -> CONFIG.get(ServerConfig.class).allowFruitStems);
         BonemealHandler.registerBehavior(Blocks.LILY_PAD, () -> new SimpleSpreadBehavior(4, 3), () -> CONFIG.get(ServerConfig.class).allowLilyPad);
         BonemealHandler.registerBehavior(Blocks.DEAD_BUSH, () -> new SimpleSpreadBehavior(4, 2), () -> CONFIG.get(ServerConfig.class).allowDeadBush);
-        BonemealHandler.registerBehavior(BlockTags.SMALL_FLOWERS, () -> new SimpleSpreadBehavior(3, 1), () -> CONFIG.get(ServerConfig.class).allowSmallFlowers);
+        BonemealHandler.registerBehavior(Set.of(Blocks.DANDELION, Blocks.POPPY, Blocks.BLUE_ORCHID, Blocks.ALLIUM, Blocks.AZURE_BLUET, Blocks.RED_TULIP, Blocks.ORANGE_TULIP, Blocks.WHITE_TULIP, Blocks.PINK_TULIP, Blocks.OXEYE_DAISY, Blocks.CORNFLOWER, Blocks.LILY_OF_THE_VALLEY), () -> new SimpleSpreadBehavior(3, 1), () -> CONFIG.get(ServerConfig.class).allowSmallFlowers);
         BonemealHandler.registerBehavior(BlockTags.CORAL_PLANTS, CoralBehavior::new, () -> CONFIG.get(ServerConfig.class).allowCorals);
         BonemealHandler.registerBehavior(Blocks.CHORUS_FLOWER, ChorusFlowerBehavior::new, () -> CONFIG.get(ServerConfig.class).allowChorus);
         BonemealHandler.registerBehavior(Blocks.CHORUS_PLANT, ChorusPlantBehavior::new, () -> CONFIG.get(ServerConfig.class).allowChorus);
         BonemealHandler.registerBehavior(Blocks.MYCELIUM, MyceliumBehavior::new, () -> CONFIG.get(ServerConfig.class).allowMycelium);
-        BonemealHandler.registerBehavior(Sets.newHashSet(Blocks.DIRT, Blocks.COARSE_DIRT, Blocks.DIRT_PATH), DirtBehavior::new, () -> CONFIG.get(ServerConfig.class).allowDirt);
+        BonemealHandler.registerBehavior(Set.of(Blocks.DIRT, Blocks.COARSE_DIRT, Blocks.DIRT_PATH), DirtBehavior::new, () -> CONFIG.get(ServerConfig.class).allowDirt);
         BonemealHandler.registerBehavior(Blocks.PODZOL, PodzolBehavior::new, () -> CONFIG.get(ServerConfig.class).allowPodzol);
+        BonemealHandler.registerBehavior(Blocks.SPORE_BLOSSOM, () -> new PopResourceBehavior(Direction.DOWN), () -> CONFIG.get(ServerConfig.class).allowSporeBlossom);
     }
 }