+++ 1.20/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.0.1-1.20.1] - 2023-08-12
+### Added
+- Added a bonemeal behavior for spore blossoms, you will receive a new spore blossom item for every bonemeal used, similar to using bonemeal on tall flowers
+### Changed
+- Growing melons and pumpkins from stems now uses much less bonemeal
+- Bonemeal no longer works on wither roses and torchflowers as there are other dedicated methods in vanilla for obtaining those
+
 ## [v8.0.0-1.20.1] - 2023-06-27
 - Ported to Minecraft 1.20.1